🚮cleanup: refactor android
