+++ android/src/main/java/com/hemangkumar/capacitorgooglemaps/CustomMarker.java
@@ -23,22 +23,24 @@ public class CustomMarker {
 
     private final MarkerOptions markerOptions = new MarkerOptions();
     private JSObject tag = new JSObject();
-    private IconDescriptor iconDescriptor;
+    private JSObject iconDescriptor;
 
     public void asyncLoadIcon(
             @NonNull FragmentActivity activity,
-            @Nullable Runnable onLoaded) {
+            @Nullable Consumer<BitmapDescriptor> consumer) {
         new AsyncIconLoader(iconDescriptor, activity)
-                .load((bitmap) -> {
-                    if (bitmap != null) {
-                        markerOptions.icon(BitmapDescriptorFactory.fromBitmap(bitmap));
-                    } else {
-                        markerOptions.icon(null);
-                    }
-                    if (onLoaded != null) {
-                        onLoaded.run();
-                    }
-                });
+            .load((bitmap) -> {
+                BitmapDescriptor bitmapDescriptor;
+                if (bitmap != null) {
+                    bitmapDescriptor = BitmapDescriptorFactory.fromBitmap(bitmap);
+                } else {
+                    bitmapDescriptor = null;
+                }
+
+                if (consumer != null) {
+                    consumer.accept(bitmapDescriptor);
+                }
+            });
     }
 
     public void updateFromJSObject(JSObject marker) {
@@ -70,25 +72,21 @@ public class CustomMarker {
 
         this.setMetadata(JSObjectDefaults.getJSObjectSafe(preferences, "metadata", new JSObject()));
 
-        JSObject jsIconDescriptor = preferences.getJSObject("icon");
-        if (jsIconDescriptor != null) {
-            iconDescriptor = new IconDescriptor(jsIconDescriptor);
-        } else {
-            iconDescriptor = null;
-        }
+        iconDescriptor = JSObjectDefaults.getJSObjectSafe(preferences, "icon", new JSObject());
     }
 
     public void addToMap(FragmentActivity activity, GoogleMap googleMap, @Nullable Consumer<Marker> consumer) {
         asyncLoadIcon(
-                activity,
-                () -> {
-                    Marker marker = googleMap.addMarker(markerOptions);
-                    marker.setTag(tag);
+            activity,
+            (BitmapDescriptor bitmapDescriptor) -> {
+                markerOptions.icon(bitmapDescriptor);
+                Marker marker = googleMap.addMarker(markerOptions);
+                marker.setTag(tag);
 
-                    if (consumer != null) {
-                        consumer.accept(marker);
-                    }
-                });
+                if (consumer != null) {
+                    consumer.accept(marker);
+                }
+            });
     }
 
     private void setMetadata(@NonNull JSObject jsObject) {
