--- android/src/main/java/com/hemangkumar/capacitorgooglemaps/ShapeTraits.java
@@ -1,62 +1,0 @@
-package com.hemangkumar.capacitorgooglemaps;
-
-public abstract class ShapeTraits {
-    boolean hasAboveMarkers() {return false;}
-    boolean hasWidth() {
-        return false;
-    }
-    boolean hasColor() {
-        return false;
-    }
-    boolean hasStrokeWidth() {
-        return false;
-    }
-
-    boolean hasStrokeColor() {
-        return false;
-    }
-
-    boolean hasFillColor() {
-        return false;
-    }
-
-    boolean hasStrokeJointType() {
-        return false;
-    }
-
-    boolean hasStrokePatterns() {
-        return false;
-    }
-
-    boolean hasPattern() {
-        return false;
-    }
-
-    boolean hasGeodesic() {
-        return false;
-    }
-
-    boolean hasHoles() {
-        return  false;
-    }
-
-    boolean hasPoints() {
-        return false;
-    }
-
-    boolean hasCenter() {
-        return false;
-    }
-
-    boolean hasRadius() {
-        return false;
-    }
-
-    boolean hasJointType() {
-        return false;
-    }
-
-    boolean hasPatternIcon() {
-        return false;
-    }
-}
